+++ app/src/features/items/index.ts
@@ -4,3 +4,4 @@ export * from './table'
 export * from './useNavigateToItem'
 export * from './editor'
 export * from './ItemIcon'
+export * from './ItemLink'
