+++ app/src/features/items/table/ItemsTable.tsx
@@ -5,7 +5,7 @@ import React, { useEffect } from 'react'
 import { selectAllTags, useStore } from '../../../data/store'
 import { Item } from '../../../data/types'
 import { renderDate } from '../../../data/util'
-import { ItemLink } from '../../../components/ItemLink'
+import { ItemLink } from '../ItemLink'
 import { ItemStatus } from '../../../components/Status'
 import { TableState, useTablesStore } from './store'
 import { SorterResult } from 'antd/lib/table/interface'
